requirements.txt
numpy
pandas
dask
dask-ml
matplotlib
scikit-learn
tensorflow
tf-keras
transformers
nltk
emoji
